refactor: keyword service 생성하여 메서드 분리
